Logo Update (#65)
* Logo Update

Upload logo was broken (it was not translating image binary properly), fixed that issue. Also added "name" header to get/set logo so company name can be stored alongside logo, and isDefault header to indicate if the returned value is the default or not. If set logo is called without name or logo, the old name or logo will not be overwritten. Updated and added tests for the logo methods. This brings java config to parity with dotnet config changes that were previously added.

* Addressed comments

Fixed some casing, adding constants

* Fixed async, style fixes

Fixed set logo to be fully asynchronous. Increased buffer size for file copy. Made various suggested style changes (comments, constants, spacing)

* Change Compose to ComposeAsync

.thenCompose should be .thenComposeAsync